src/main/python/requirements.txt
matplotlib==2.8.1
pandas==1.0.4
pillow==7.2.0
